requirements.txt
streamlit==1.47.1
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
matplotlib==3.9.2
seaborn==0.13.2
